Rename saga 'machine core' naming to 'core' in javadoc/kdoc and example package
